+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The `pycldf` package adheres to [Semantic Versioning](http://semver.org/spec/v2.
 ## unreleased
 
 - Support certain GitHub URLs as dataset locators.
+- Pragmatic support for large media files.
 
 
 ## [1.39.0] - 2024-09-09